поменял на long от переполнения

+++ task_2_1_2/src/main/java/worker/Worker.java
@@ -63,7 +63,7 @@ public class Worker {
      */
     static boolean isComposite(int n) {
         if (n < 4) return false;
-        for (int i = 2; i * i <= n; i++) {
+        for (long i = 2; i * i <= (long) n; i++) {
             if (n % i == 0) return true;
         }
         return false;
